Outrelix/Outrelix-Scraper/requirements.txt
requests
pandas
openpyxl
python-dotenv
tenacity
phonenumbers
rapidfuzz
tqdm
gspread
gspread_dataframe
oauth2client
httpx[socks]
schedule
fastapi==0.111.0
uvicorn[standard]==0.30.0
PyJWT==2.9.0
beautifulsoup4
lxml
regex
transformers
torch
playwright
sentence-transformers
scikit-learn
dnspython
aiosmtplib

